Add files via upload
Repository Name: Introduction to Statistics

Description:
Welcome to the "Introduction to Statistics " repository! 📊🐍 In this comprehensive guide, we cover the fundamental concepts of statistics, types of data, and types of statistics, and provide an introduction to descriptive statistics using Python.

Key Topics Covered:

Types of Data:

Learn about different types of data, including categorical and numerical.
Understand the importance of choosing the right data type for your analysis.
Types of Statistics:

Explore the distinctions between descriptive and inferential statistics.
Gain insights into how descriptive statistics summarize and describe features of a dataset.
Descriptive Statistics with Python.


Who Is This For?

Students and beginners in statistics and data analysis.
Python enthusiasts looking to apply statistical concepts in their projects.
Is anyone interested in gaining a solid understanding of descriptive statistics?
Get Started:

Explore the README for an overview of the repository.
Navigate to the "Descriptive_Statistics" folder for Python code examples.
Engage in discussions, ask questions, and share your experiences.
Join us in this statistical exploration with Python! 📉🐍